pyproject.toml
[project]
name = "notisense-api"
version = "0.1.0"
description = ""
authors = [
    {name = "AFOLABI GMATT MATTHEW",email = "gmattworld@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "python-decouple (>=3.8,<4.0)",
    "fastapi-mail (>=1.5.0,<2.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "greenlet (>=3.2.4,<4.0.0)",
    "mangum (>=0.19.0,<0.20.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)"
]

[tool.poetry]
packages = [{include = "notisense_api", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
